Refactor /students routes as controller and query
This commit moves /students and /students/:id from app.js
into a dedicated students conroller. It also extracts the
data-fetching functions from the controller to a query layer.
