+++ app.js
@@ -7,7 +7,7 @@ const express = require('express');
 const cors = require('cors');
 
 // Import OUR stuff (our files, our components)
-const studentsData = require('./studentsData.json');
+const studentsController = require('./controllers/studentsController');
 
 // Init express application
 const app = express();
@@ -17,6 +17,9 @@ const app = express();
 // the final route handler function
 app.use(cors());
 
+// Controllers
+app.use('/students', studentsController);
+
 // Define our routes
 
 // Healthcheck route
@@ -24,32 +27,6 @@ app.get('/', (request, response) => {
   response.status(200).json({ data: 'Service is running' });
 });
 
-app.get('/students', (request, response) => {
-  try {
-    const { students } = studentsData;
-    response.status(200).json({ data: students });
-  } catch (err) {
-    response.status(500).json({ error: err.message });
-  }
-});
 
-app.get('/students/:id', (request, response) => {
-  try {
-    const { id } = request.params;
-    const { students } = studentsData;
-
-    const student = students.find((el) => el.id === id);
-    if (student) {
-      // return 200
-      return response.status(200).json({ data: student });
-    }
-    // return 404
-    response
-      .status(404)
-      .json({ error: `Could not find student with id ${id}` });
-  } catch (err) {
-    response.status(500).json({ error: err.message });
-  }
-});
 
 module.exports = app;
