+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "mike.js",
   "scripts": {
+    "start": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
